+++ guillotina/tests/cache/test_cache_pubsub.py
@@ -18,7 +18,12 @@ DEFAULT_SETTINGS = {
         'guillotina.contrib.redis',
         'guillotina.contrib.cache',
         'guillotina.contrib.pubsub',
-    ]
+    ],
+    'cache': {
+        'updates_channel': 'guillotina',
+        'driver': 'guillotina.contrib.redis'
+    }
+
 }